C++ Warning Reduction
Pivotal Tracker: #72898288

Additional author: Feng, Xixuan (Aaron) <xfeng@gopivotal.com>

Changes:
- add many static_cast
- fix mlogregr_result to use bigint for num_processed

+++ src/modules/convex/utils_regularization.cpp
@@ -131,8 +131,8 @@ AnyType utils_var_scales_final::run (AnyType& args)
 
     if (state.numRows == 0) return Null();
 
-    state.mean /= state.numRows;
-    state.std /= state.numRows;
+    state.mean /= static_cast<double>(state.numRows);
+    state.std /= static_cast<double>(state.numRows);
     for (uint32_t i = 0; i < state.dimension; i++)
         state.std(i) = sqrt(state.std(i) - state.mean(i) * state.mean(i));
